put all models under mudslide.models
